testcontainers-vitalmagic

+++ src/main/java/com/lps/vitalMagic/shake/infrastructure/persistence/projection/ShakeAttributeProjection.java
@@ -1,9 +1,11 @@
 package com.lps.vitalMagic.shake.infrastructure.persistence.projection;
 
+import java.math.BigDecimal;
+
 public record ShakeAttributeProjection(
         Long shakeId,
         Long attributeId,
         String attributeName,
-        Integer total
+        BigDecimal total
 ) {
 }
